requirements.txt
discord.py>=2.3.2
simpleeval>=0.9.13
python-dotenv>=1.0.1
